# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saiemaa's Restaurant API
 * OpenAPI spec version: 0.2.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  name: string;
  phone: string;
  date: string;
  time: string;
  guests: number;
  requests?: string;
  status: BookingStatus;
  tableId?: number;
  notes?: string;
  createdAt: string;
}