# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saiemaa's Restaurant API
 * OpenAPI spec version: 0.2.0
 */
import type { ContactStatus } from "./contactStatus";

export interface Contact {
  id: number;
  name: string;
  phone: string;
  email?: string;
  message?: string;
  status: ContactStatus;
  notes?: string;
  createdAt: string;
}