# lib/api-zod/src/generated/types/createTableInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saiemaa's Restaurant API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateTableInputStatus } from "./createTableInputStatus";

export interface CreateTableInput {
  number: number;
  capacity: number;
  location?: string;
  status?: CreateTableInputStatus;
}