# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saiemaa's Restaurant API
 * OpenAPI spec version: 0.2.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  tableId: number;
  tableNumber?: number;
  customerName?: string;
  customerPhone?: string;
  items: string;
  totalAmount?: string;
  status: OrderStatus;
  notes?: string;
  createdAt: string;
  updatedAt?: string;
}