# lib/api-zod/src/generated/types/updateContactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saiemaa's Restaurant API
 * OpenAPI spec version: 0.2.0
 */
import type { UpdateContactInputStatus } from "./updateContactInputStatus";

export interface UpdateContactInput {
  status?: UpdateContactInputStatus;
  notes?: string;
}