# lib/api-zod/src/generated/types/updateOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saiemaa's Restaurant API
 * OpenAPI spec version: 0.2.0
 */
import type { UpdateOrderInputStatus } from "./updateOrderInputStatus";

export interface UpdateOrderInput {
  status?: UpdateOrderInputStatus;
  notes?: string;
  totalAmount?: string;
}